Added abstract classes, split services in different files.
